// frontend/src/gen/types/Frequency.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export const frequencyEnum = {
  weekly: "weekly",
  fortnightly: "fortnightly",
  every_3_weeks: "every_3_weeks",
  every_4_weeks: "every_4_weeks",
} as const;

export type FrequencyEnumKey =
  (typeof frequencyEnum)[keyof typeof frequencyEnum];

/**
 * Frequency
 */
export type Frequency = FrequencyEnumKey;