// frontend/src/gen/types/SeriesOut.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { Frequency } from "./Frequency.ts";

/**
 * SeriesOut
 */
export type SeriesOut = {
  /**
   * @type integer
   */
  id: number;
  /**
   * @type integer
   */
  client_id: number;
  /**
   * @type string
   */
  client_name: string;
  /**
   * @type string
   */
  frequency: Frequency;
  /**
   * @type string, date
   */
  start_date: string;
  default_time: string | null;
  /**
   * @type integer
   */
  duration_minutes: number;
  end_date: string | null;
};